updated the javascript to replace jquery event hooking/triggering with regular javascript functions.  Talked it over with Russ a few weeks ago, it worked out pretty well.
Now we keep an object that is the current command set, and we have a function to merge in new command sets.  This gives essentially a command context we can modify based on different input to enable context-sensitive UI.  When the mute button is hit, we assert the "moving player" command context, which overwrites a property in the current command set to do something a little different.  When mute is hit again, we leave the "moving player" command context by asserting the "standard" command context, which contains our default definitions for what to do for certain input.

Also added keyboard support, mapping keypresses to IR input strings.  This makes testing/dev a LOT easier.